0.1.4 with fix for NumberFormatException while parsing integer literals in SQL statements.
 Instead of Long, Use BigInt for retaining integer literals inside IntLiteral.

+++ src/main/scala/parser.scala
@@ -254,7 +254,7 @@ class SQLParser extends StandardTokenParsers {
   override def stringLit: Parser[String] = inParenthesis(super.stringLit)
 
   def literal: Parser[SqlExpr] = inParenthesis(
-    numericLit ^^ { case i => IntLiteral(i.toInt) } |
+    numericLit ^^ { case i => IntLiteral(BigInt(i)) } |
     floatLit ^^ { case f => FloatLiteral(f.toDouble) } |
     stringLit ^^ { case s => StringLiteral(s) } |
     "null".ignoreCase ^^ (_ => NullLiteral()) |
